experiments/experiment03/blog_sonnet_codeoath/requirements.txt
flask>=3.0
markdown>=3.6
pyyaml>=6.0
